Remove the Popover funnel, it was a bit silly
